better folder structure and schema changes
