+++ packages/ctfv-backend/drizzle/0001_free_sebastian_shaw.sql
@@ -1,0 +1,1 @@
+ALTER TABLE `challenges` ADD `author` text NOT NULL;
