+++ packages/ctfv-backend/drizzle/meta/_journal.json
@@ -8,6 +8,13 @@
       "when": 1726770450832,
       "tag": "0000_watery_miss_america",
       "breakpoints": true
+    },
+    {
+      "idx": 1,
+      "version": "6",
+      "when": 1726811911644,
+      "tag": "0001_free_sebastian_shaw",
+      "breakpoints": true
     }
   ]
 }
